+++ after/plugin/completion.lua
@@ -26,13 +26,9 @@ cmp.setup {
   },
 
   sources = {
-    { name = "nvim_lsp_signature_help" },
     { name = "nvim_lsp" },
     { name = "nvim_lua" },
-    { name = "cmp_tabnine" },
-    { name = "path" },
     { name = "luasnip" },
-    { name = "buffer", keyword_length = 5, max_item_count = 5 },
   },
 
   sorting = {