+++ after/plugin/vim-wiki.lua
@@ -1,7 +1,0 @@
-vim.g.vimwiki_list = {
-  {
-    ext = 'md',
-    path = '~/Documents/Google Drive/vim-wiki',
-    syntax = 'markdown',
-  }
-}